--- lib/whatsapp_sender_flutter_response_message.dart
@@ -1,8 +1,0 @@
-class WhatsAppSenderFlutterResponseMessageValues {
-  static const String initializing = "initializing";
-  static const String errorOnLaunch = "error on launch";
-  static const String scanQrCode = "scan qr code";
-  static const String sending = "sending";
-  static const String done = "done";
-  static const String qrCodeExpirated = "qr code expirated";
-}
